backend/pyproject.toml
[tool.poetry]
name = "workflow-backend"
version = "0.1.0"
description = "Production-grade workflow engine with FastAPI and LangGraph"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
supabase = "^2.3.0"
langgraph = "^0.6.6"  # Updated to latest version
langchain-core = "^0.3.74"
openai = "^1.10.0"
python-dotenv = "^1.0.0"
jmespath = "^1.0.1"
urllib3 = "<2.0.0"
langchain = "^0.3.27"
langchain-community = "^0.3.27"
typing-extensions = "^4.14.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
black = "^24.0.0"

[tool.poetry.scripts]
start = "app.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=app"
testpaths = ["tests"]